+++ app/routes/admin.py
@@ -44,7 +44,7 @@ async def admin_update_post_by_id(post_id: int,
                                   created_at: datetime = None,
                                   updated_at: datetime = None,
                                   tags: str = None,
-                                  rating: int = None
+                                  rating: float = None
                                                         ) -> Post | None:
     """
     Update a post by id, the function works only for users with administrator rights.
